refactor: extract game session schedule schemas and types into their own files
