+++ packages/shared/src/schemas/index.ts
@@ -2,6 +2,7 @@ export * from "./auth"
 export * from "./faq"
 export * from "./footer"
 export * from "./game-session"
+export * from "./game-session-schedule"
 export * from "./link"
 export * from "./media"
 export * from "./navbar"
