+++ packages/shared/src/types/game-session-schedule.ts
@@ -1,0 +1,12 @@
+import type z from "zod"
+import type {
+  CreateGameSessionScheduleRequestSchema,
+  UpdateGameSessionScheduleRequestSchema,
+} from "../schemas"
+
+export type CreateGameSessionScheduleRequest = z.infer<
+  typeof CreateGameSessionScheduleRequestSchema
+>
+export type UpdateGameSessionScheduleRequest = z.infer<
+  typeof UpdateGameSessionScheduleRequestSchema
+>
