+++ packages/shared/src/types/game-session.ts
@@ -1,14 +1,4 @@
 import type z from "zod"
-import type {
-  CreateGameSessionScheduleRequestSchema,
-  GameSessionScheduleSchema,
-  UpdateGameSessionScheduleRequestSchema,
-} from "../schemas"
+import type { GameSessionScheduleSchema } from "../schemas"
 
 export type GameSessionSchedule = z.infer<typeof GameSessionScheduleSchema>
-export type CreateGameSessionScheduleRequest = z.infer<
-  typeof CreateGameSessionScheduleRequestSchema
->
-export type UpdateGameSessionScheduleRequest = z.infer<
-  typeof UpdateGameSessionScheduleRequestSchema
->
