+++ packages/shared/src/types/index.ts
@@ -2,6 +2,7 @@ export * from "./auth"
 export * from "./collections"
 export * from "./enums"
 export * from "./game-session"
+export * from "./game-session-schedule"
 export * from "./payload"
 export * from "./query"
 export * from "./richtext"
